split up /ghauth with /ghverify
